Update for multi-column support
BetaFlight 4.4 is using a 4 page font. The in goggle font file format will change from a single character wide by 256/512 character high font file; to a 1-4 character wide, 256 character high font file.. This tool has now been updated to support the multiple column file format. Currently, you need to use the `-xcols` argument to generate the multi column font. But when the change of the font file is in the goggles firmware, I will updated this tool to default to multi-column, and have a legacy, single column argument.